# Add support for faster uploads using Azure Storage Data Movement Library